# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News API",
        "description": "API для керування новинами",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8082/api",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/comments": {
            "post": {
                "summary": "Додати коментар",
                "operationId": "478e0f4964ebbe8d3a22df34f4fcddab",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "news_id",
                                    "content"
                                ],
                                "properties": {
                                    "news_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Гарна новина!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Коментар додано"
                    },
                    "422": {
                        "description": "Помилка валідації"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/comments/{id}": {
            "put": {
                "summary": "Оновити коментар",
                "operationId": "0904942451844b2582af15f7c38988b4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "content"
                                ],
                                "properties": {
                                    "content": {
                                        "type": "string",
                                        "example": "Оновлений коментар"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Коментар оновлено"
                    },
                    "404": {
                        "description": "Коментар не знайдено"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "summary": "Видалити коментар",
                "operationId": "dc71fcedd4e6d8fe9a74a965257cc90b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Коментар видалено"
                    },
                    "404": {
                        "description": "Коментар не знайдено"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/my-comments": {
            "get": {
                "summary": "Мої коментарі",
                "operationId": "eaef051645d734e86e47d6c55a523ef5",
                "responses": {
                    "200": {
                        "description": "Список коментарів"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}